refactor: align serverOnly mode with fullstack server architecture
- Revert generateServerEntry to original shape (export createApp + _fns_*)
- Delete template.ts and create-server-webpack-config.ts
- Unify createWebpackConfig with serverOnlyEntry param
- Align dev bootstrap: both modes use createApp() + serve()
- Update E2E fixture to use isomorphic createApp bootstrap
- serverOnly output is now isomorphic with fullstack server-side
