+++ packages/build-tools/src/types.ts
@@ -1,11 +1,6 @@
 /** Configuration for the generated server entry. */
 export interface ServerEntryConfig {
   /**
-   * Runtime adapter module that exports `createFetchHandler(app)`.
-   * Default: "@evjs/runtime/server/ecma"
-   */
-  runner?: string;
-  /**
    * Middleware module paths to auto-register in the server entry.
    */
   middleware?: string[];
@@ -19,14 +14,10 @@ export interface TransformOptions {
   rootContext: string;
   /** Whether this is a server-side build. */
   isServer: boolean;
-  /**
-   * Use human-readable function IDs (`relativePath#exportName`)
-   * instead of hashed IDs. Useful for FaaS mode where there's no
-   * client bundle and readable IDs aid debugging.
-   */
+  /** Whether to use readable function IDs (e.g. "src/api/users.server#getUsers"). */
   readableIds?: boolean;
   /**
-   * If true, process the file even if it lacks the "use server" directive.
+   * When true, skip the "use server" directive check.
    * Useful in FaaS mode.
    */
   ignoreDirective?: boolean;
@@ -52,10 +43,6 @@ export const RUNTIME = {
   serverModule: "@evjs/runtime/server/register",
   /** Module path for the server app factory (Hono app + server function handler). */
   appModule: "@evjs/runtime/server",
-  /** Module path for the ECMA server environment fetch handler. */
-  ecmaModule: "@evjs/runtime/server/ecma",
-  /** Module path for the Node server environment runner. */
-  nodeModule: "@evjs/runtime/server/node",
   /** Module path for client-side transport stubs. */
   clientTransportModule: "@evjs/runtime/client/transport",
   /** Server function registration call name. */
