+++ packages/build-tools/tests/entry.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from "vitest";
 import { generateServerEntry } from "../src/entry.js";
 
 describe("generateServerEntry", () => {
-  it("generates an isomorphic entry with user functions, createApp, and a default fetch handler", () => {
+  it("generates entry with user functions and createApp", () => {
     const result = generateServerEntry(undefined, [
       "/project/src/api/users.server.ts",
     ]);
@@ -10,10 +10,8 @@ describe("generateServerEntry", () => {
     expect(result).toContain("import * as _fns_0");
     expect(result).toContain("export { _fns_0 }");
     expect(result).toContain(
-      'import { createApp } from "@evjs/runtime/server"',
+      'export { createApp } from "@evjs/runtime/server"',
     );
-    expect(result).toContain("export { createApp }");
-    expect(result).toContain("export default createFetchHandler(app)");
   });
 
   it("imports and re-exports all server modules", () => {
@@ -48,41 +46,7 @@ describe("generateServerEntry", () => {
     const result = generateServerEntry(undefined, []);
 
     expect(result).not.toContain("import * as _fns");
-    // Still creates app and exports handler
+    // Still exports createApp for the adapter layer
     expect(result).toContain("export { createApp }");
-    expect(result).toContain("export default createFetchHandler(app)");
-  });
-
-  it("passes endpoint to createApp when provided", () => {
-    const result = generateServerEntry(
-      undefined,
-      ["/project/src/api/hello.server.ts"],
-      "/rpc/v1",
-    );
-
-    expect(result).toContain('endpoint: "/rpc/v1"');
-  });
-
-  it("uses default createApp call when no endpoint is provided", () => {
-    const result = generateServerEntry(undefined, [
-      "/project/src/api/hello.server.ts",
-    ]);
-
-    expect(result).toContain("const app = createApp()");
-  });
-
-  it("includes middleware imports before server modules", () => {
-    const result = generateServerEntry(
-      {
-        middleware: ['import "dotenv/config";'],
-      },
-      ["/project/src/api/hello.server.ts"],
-    );
-
-    expect(result).toContain('import "dotenv/config"');
-    // Middleware should come before module imports
-    const dotenvIndex = result.indexOf('import "dotenv/config"');
-    const fnsIndex = result.indexOf("import * as _fns_0");
-    expect(dotenvIndex).toBeLessThan(fnsIndex);
   });
 });
